Move CompositeFetcher to a separate file
